requirements.txt
asgiref==3.6.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.0.1
convertdate==2.4.0
Django==3.2.16
django-crispy-forms==1.14.0
djangorestframework==3.14.0
docutils==0.19
hijri-converter==2.2.4
holidays==0.19
idna==3.4
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
korean-lunar-calendar==0.3.1
Pillow==9.4.0
Pygments==2.14.0
PyMeeus==0.5.12
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pywin32==305
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
wincertstore==0.2
